use www/settings.json to control upload method
